requirements.txt
pandas
numpy
python-dotenv
pydantic
scikit-learn
openai
